rename: add renaming and dep analysis for class morphisms
The basic idea is that we bring morphisms into the dependency graph
(under a phony name) making class C depend on every morphism from C to
any other class. Of course, the morphism depends on the class as well,
so they'll be checked in the same recursive group. This might also cause
different classes to have to share a group (C -> D, D -> C).

We do this since at the time of tcInstDecls1, we need to expand the
context of instances, and we better have all relevant morphisms
in-scope. This setup guarantees that, if class C is in-scope, then its
entire deductive closure is as well.